@@ -1,47 +1,49 @@
 name: allo
 description: Chat app
 publish_to: 'none'
-version: 1.3.4+228
+version: 1.3.4+274
 
 environment:
-  sdk: ">=2.15.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   animations: ^2.0.2
   awesome_notifications: #^0.6.21 
     git:
-      ref: "9304833"
+      ref: "94854e8"
       url: https://github.com/rafaelsetragni/awesome_notifications
   cached_network_image: ^3.2.0
-  cloud_firestore: ^3.1.6
-  device_info_plus: ^3.2.2
-  dynamic_color: ^1.1.2
-  firebase_auth: ^3.3.5
-  firebase_core: ^1.12.0
-  firebase_messaging: ^11.2.6
-  firebase_remote_config: ^2.0.0
-  firebase_storage: ^10.2.6
+  cloud_firestore: ^3.1.15
+  device_info_plus: ^3.2.3
+  dynamic_color: ^1.2.2
+  firebase_auth: ^3.3.18
+  firebase_core: ^1.17.0
+  firebase_messaging: ^11.4.0
+  firebase_remote_config: ^2.0.7
+  firebase_storage: ^10.2.16
   firebase_image:
     path: ./packages/firebase_image
   flutter:
     sdk: flutter
-  flutter_hooks: ^0.18.0
+  flutter_hooks: ^0.18.4
   flutter_linkify: ^5.0.2
   flutter_localizations:
     sdk: flutter
   flutter_platform_widgets: ^1.20.0
-  hooks_riverpod: ^1.0.1
+  hooks_riverpod: ^2.0.0-dev.7
   http: ^0.13.4
-  image_cropper: ^1.4.1
-  image_picker: ^0.8.4+6
+  image_cropper: ^2.0.2
+  image_picker: ^0.8.5+3
   # in_app_update: ^2.0.0 // Some bugs do not permit for this to run. We will remove it temporarily as it is not used either.
   intl: ^0.17.0
-  package_info_plus: ^1.4.0
-  shared_preferences: ^2.0.13
-  url_launcher: ^6.0.18
+  package_info_plus: ^1.4.2
+  shared_preferences: ^2.0.15
+  url_launcher: ^6.1.2
+  flutter_native_splash: ^2.1.5
+  material_color_utilities: ^0.1.4
 
 dev_dependencies:
-  flutter_lints: ^1.0.4
+  flutter_lints: ^2.0.1
   flutter_test:
     sdk: flutter
 
