@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 class UserPreviewPage extends HookConsumerWidget {
-  const UserPreviewPage({Key? key}) : super(key: key);
+  const UserPreviewPage({super.key});
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     return Scaffold(
