@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 
 class AccountSettings extends HookWidget {
-  const AccountSettings({Key? key}) : super(key: key);
+  const AccountSettings({super.key});
   @override
   Widget build(BuildContext context) {
     final locales = S.of(context);
