@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 
 class AccountInfo extends HookWidget {
-  const AccountInfo({Key? key}) : super(key: key);
+  const AccountInfo({super.key});
 
   @override
   Widget build(BuildContext context) {
