@@ -7,7 +7,7 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
 
 class CreateChat extends HookWidget {
-  const CreateChat({Key? key}) : super(key: key);
+  const CreateChat({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -15,33 +15,60 @@ class CreateChat extends HookWidget {
     final loading = useState(false);
     final error = useState<String?>(null);
     final usernameController = useTextEditingController();
+    final found = useState(<String>[]);
     return Scaffold(
       appBar: AppBar(
         title: Text(locales.createNewChat),
       ),
-      body: FutureView<Map<String, dynamic>?>(
+      body: FutureView<Map<String, String>>(
         future: Core.general.user.getUsernamePairs(),
         success: (context, data) {
           return Padding(
             padding: const EdgeInsets.all(20),
             child: Column(
               children: [
-                const Text(
-                  'Type below the username of the user you want to create a chat with.',
-                ),
-                const Space(2),
                 TextFormField(
                   decoration: InputDecoration(
                     contentPadding: const EdgeInsets.all(10),
                     errorText: error.value,
                     errorStyle: const TextStyle(fontSize: 14),
                     labelText: locales.username,
-                    border: OutlineInputBorder(
-                      borderRadius: BorderRadius.circular(100),
-                    ),
+                    border: const OutlineInputBorder(),
                   ),
+                  onChanged: (value) {
+                    final prov = <String>[];
+                    for (final element in data.keys) {
+                      if (element.toString().contains(value)) {
+                        prov.add(element);
+                      }
+                    }
+                    found.value = prov;
+                  },
                   controller: usernameController,
                 ),
+                const Space(2),
+                if (found.value.isEmpty) ...[
+                  const Expanded(
+                    flex: 5,
+                    child: Align(
+                      child: Text(
+                        'No items found.',
+                      ),
+                    ),
+                  )
+                ] else ...[
+                  Expanded(
+                    flex: 10,
+                    child: ListView.builder(
+                      itemCount: found.value.length,
+                      itemBuilder: (_, i) {
+                        return ListTile(
+                          title: Text(found.value[i]),
+                        );
+                      },
+                    ),
+                  ),
+                ],
                 Expanded(
                   child: Align(
                     alignment: Alignment.bottomCenter,
@@ -83,7 +110,7 @@ class CreateChat extends HookWidget {
                         if (usernameController.text.isEmpty) {
                           error.value = context.locale.errorFieldEmpty;
                         }
-                        if (!data!.containsKey(usernameController.text)) {
+                        if (!data.containsKey(usernameController.text)) {
                           error.value = 'This username does not exist.';
                         } else {
                           showPlatformDialog(
