@@ -1,26 +1,23 @@
-import 'package:allo/components/oobe_page.dart';
+import 'package:allo/components/setup_page.dart';
 import 'package:allo/generated/l10n.dart';
+import 'package:allo/logic/core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 
-import '../home/settings/debug/debug.dart';
 import 'login.dart';
 
 class Setup extends HookWidget {
-  const Setup({Key? key}) : super(key: key);
+  const Setup({super.key});
   @override
   Widget build(BuildContext context) {
     final locales = S.of(context);
     return SetupPage(
       icon: Icons.emoji_people,
-      title: locales.setupWelcomeScreenTitle,
-      subtitle: locales.setupWelcomeScreenDescription,
-      body: const [],
-      debug: const C(),
+      title: Text(locales.setupWelcomeScreenTitle),
+      subtitle: Text(locales.setupWelcomeScreenDescription),
       action: () async {
-        return true;
+        Core.navigation.push(route: const Login());
       },
-      nextRoute: const Login(),
     );
   }
 }
