@@ -1,23 +1,22 @@
-import 'package:allo/components/oobe_page.dart';
+import 'package:allo/components/setup_page.dart';
 import 'package:allo/interface/home/tabbed_navigator.dart';
 import 'package:allo/logic/core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 
 class SetupDone extends HookWidget {
-  const SetupDone({Key? key}) : super(key: key);
+  const SetupDone({super.key});
   @override
   Widget build(BuildContext context) {
+    void onSubmit() {
+      Core.navigation.push(route: const TabbedNavigator());
+    }
+
     return SetupPage(
       icon: Icons.check_circle,
-      title: context.locale.finishScreenTitle,
-      subtitle: context.locale.finishScreenDescription,
-      body: const [],
-      action: () async {
-        return true;
-      },
-      isRoutePermanent: true,
-      nextRoute: TabbedNavigator(),
+      title: Text(context.locale.finishScreenTitle),
+      subtitle: Text(context.locale.finishScreenDescription),
+      action: onSubmit,
     );
   }
 }
