@@ -1,22 +1,24 @@
-import 'package:allo/components/oobe_page.dart';
 import 'package:allo/components/person_picture.dart';
-import 'package:allo/generated/l10n.dart';
+import 'package:allo/components/setup_page.dart';
 import 'package:allo/interface/login/new/setup_personalize.dart';
 import 'package:allo/logic/core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 
 class SetupProfilePicture extends HookWidget {
-  const SetupProfilePicture({Key? key}) : super(key: key);
+  const SetupProfilePicture({super.key});
   @override
   Widget build(BuildContext context) {
-    final locales = S.of(context);
     final percentage = useState(0.0);
     final loaded = useState(false);
+    void onSubmit() {
+      Core.navigation.push(route: const SetupPersonalize());
+    }
+
     return SetupPage(
       icon: Icons.landscape,
-      title: context.locale.setupProfilePictureScreenTitle,
-      subtitle: context.locale.setupProfilePictureScreenDescription,
+      title: Text(context.locale.setupProfilePictureScreenTitle),
+      subtitle: Text(context.locale.setupProfilePictureScreenDescription),
       body: [
         Column(
           children: [
@@ -42,7 +44,7 @@ class SetupProfilePicture extends HookWidget {
               ),
             ),
             ListTile(
-              title: Text(locales.uploadPicture),
+              title: Text(context.locale.uploadPicture),
               onTap: () async => await Core.auth.user.updateProfilePicture(
                 loaded: loaded,
                 percentage: percentage,
@@ -53,8 +55,7 @@ class SetupProfilePicture extends HookWidget {
           ],
         ),
       ],
-      action: () async => true,
-      nextRoute: const SetupPersonalize(),
+      action: onSubmit,
     );
   }
 }
