@@ -1,18 +1,18 @@
-import 'package:allo/generated/l10n.dart';
+import 'package:allo/components/setup_page.dart';
 import 'package:allo/interface/login/new/setup_pfp.dart';
-import 'package:allo/logic/client/extensions.dart';
+import 'package:allo/logic/core.dart';
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:allo/components/oobe_page.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
 
+const _kDebugBypassVerification = kDebugMode ? true : false;
+
 class SetupVerification extends HookWidget {
-  const SetupVerification({Key? key}) : super(key: key);
+  const SetupVerification({super.key});
   @override
   Widget build(BuildContext context) {
-    final locales = S.of(context);
-
     useEffect(() {
       Future.microtask(
         () async =>
@@ -22,20 +22,19 @@ class SetupVerification extends HookWidget {
     });
     return SetupPage(
       icon: Icons.email,
-      title: context.locale.setupVerificationScreenTitle,
-      subtitle: context.locale.setupVerificationScreenDescription,
-      body: const [],
+      title: Text(context.locale.setupVerificationScreenTitle),
+      subtitle: Text(context.locale.setupVerificationScreenDescription),
       action: () async {
         await FirebaseAuth.instance.currentUser?.reload();
         final verified = FirebaseAuth.instance.currentUser!.emailVerified;
-        if (verified) {
-          return true;
+        if (verified || _kDebugBypassVerification) {
+          Core.navigation.push(route: const SetupProfilePicture());
         } else {
           await showPlatformDialog(
             context: context,
             builder: (context) => AlertDialog(
               title: Text(context.locale.error),
-              content: Text(locales.errorVerificationLinkNotAccessed),
+              content: Text(context.locale.errorVerificationLinkNotAccessed),
               actions: [
                 TextButton(
                   onPressed: () => Navigator.pop(context),
@@ -44,10 +43,8 @@ class SetupVerification extends HookWidget {
               ],
             ),
           );
-          return false;
         }
       },
-      nextRoute: const SetupProfilePicture(),
     );
   }
 }
