@@ -4,8 +4,7 @@ enum Direction { vertical, horizontal }
 
 /// A widget to use instead of bottom padding when you want to space components.
 class Space extends StatelessWidget {
-  const Space(this.factor, {this.direction = Direction.vertical, Key? key})
-      : super(key: key);
+  const Space(this.factor, {this.direction = Direction.vertical, super.key});
   final double factor;
   final Direction direction;
   @override
