@@ -1,51 +1,91 @@
-import 'package:allo/components/oobe_page.dart';
+import 'package:allo/components/setup_page.dart';
 import 'package:allo/components/space.dart';
 import 'package:allo/generated/l10n.dart';
 import 'package:allo/interface/home/tabbed_navigator.dart';
+import 'package:allo/logic/backend/setup/login.dart';
 import 'package:allo/logic/core.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-class EnterPassword extends HookWidget {
-  final String email;
-  const EnterPassword({required this.email, Key? key}) : super(key: key);
+class EnterPassword extends HookConsumerWidget {
+  const EnterPassword({super.key});
   @override
-  Widget build(BuildContext context) {
-    final error = useState('');
+  Widget build(BuildContext context, WidgetRef ref) {
+    final error = useState<String?>(null);
     final obscure = useState(true);
+    final focusNode = useFocusNode();
+    final login = ref.watch(loginState.notifier);
+    final state = ref.watch(loginState);
     final controller = useTextEditingController();
     final locales = S.of(context);
+
+    void onSubmit() async {
+      try {
+        await login.login(controller.text);
+        Core.navigation.pushPermanent(route: const TabbedNavigator());
+      } on FirebaseAuthException catch (e) {
+        switch (e.code) {
+          case 'user-disabled':
+            error.value = locales.errorUserDisabled;
+            break;
+          case 'wrong-password':
+            error.value = locales.errorWrongPassword;
+            break;
+          case 'too-many-requests':
+            error.value = locales.errorTooManyRequests;
+            break;
+          default:
+            error.value = locales.errorUnknown;
+            break;
+        }
+        focusNode.requestFocus();
+      } catch (e) {
+        error.value = e.toString();
+        focusNode.requestFocus();
+      }
+    }
+
     return SetupPage(
-      alignment: CrossAxisAlignment.start,
       icon: Icons.password,
-      title: '${locales.welcomeBack}.',
-      subtitle: locales.enterPasswordDescription,
+      title: Text('${locales.welcomeBack}.'),
+      subtitle: Text(locales.enterPasswordDescription),
       body: [
         TextFormField(
           decoration: InputDecoration(
             contentPadding: const EdgeInsets.all(10),
-            errorText: error.value == '' ? null : error.value,
+            errorText: error.value,
             errorStyle: const TextStyle(fontSize: 14),
             labelText: locales.password,
             border: const OutlineInputBorder(),
-            suffix: InkWell(
-              onTap: () {
-                if (obscure.value) {
-                  obscure.value = false;
-                } else {
-                  obscure.value = true;
-                }
-              },
-              child: Padding(
-                padding: const EdgeInsets.only(right: 10.0),
-                child: Icon(
-                  obscure.value ? Icons.visibility_off : Icons.visibility,
+            suffix: Padding(
+              padding: const EdgeInsets.all(5),
+              child: SizedBox(
+                height: 20,
+                width: 20,
+                child: IconButton(
+                  iconSize: 25,
+                  // ignore: use_named_constants
+                  padding: const EdgeInsets.all(0),
+                  color: context.colorScheme.primary,
+                  icon: obscure.value
+                      ? const Icon(
+                          Icons.visibility,
+                        )
+                      : const Icon(
+                          Icons.visibility_off,
+                        ),
+                  onPressed: () => obscure.value = !obscure.value,
                 ),
               ),
             ),
           ),
+          autofocus: true,
+          focusNode: focusNode,
           controller: controller,
-          obscureText: true,
+          obscureText: obscure.value,
+          onFieldSubmitted: (string) async => onSubmit(),
         ),
         const Space(2),
         TextButton(
@@ -54,21 +94,16 @@ class EnterPassword extends HookWidget {
             alignment: Alignment.topLeft,
           ),
           onPressed: () {
-            Core.auth.sendPasswordResetEmail(email: email, context: context);
+            if (state != null) {
+              Core.auth.sendPasswordResetEmail(email: state, context: context);
+            } else {
+              throw Exception('There is no email in state.');
+            }
           },
           child: Text(locales.forgotPassword),
         )
       ],
-      action: () async {
-        return await Core.auth.signIn(
-          email: email,
-          password: controller.text,
-          context: context,
-          error: error,
-        );
-      },
-      nextRoute: TabbedNavigator(),
-      isRoutePermanent: true,
+      action: onSubmit,
     );
   }
 }
