@@ -1,6 +1,5 @@
-import 'package:allo/components/oobe_page.dart';
+import 'package:allo/components/setup_page.dart';
 import 'package:allo/generated/l10n.dart';
-import 'package:allo/interface/login/new/setup_verification.dart';
 import 'package:allo/logic/core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
@@ -10,44 +9,50 @@ class SetupPassword extends HookWidget {
     required this.displayName,
     required this.username,
     required this.email,
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
   final String displayName;
   final String username;
   final String email;
   @override
   Widget build(BuildContext context) {
-    final error = useState('');
+    final error = useState<String?>(null);
     final locales = S.of(context);
     final obscure = useState(true);
     final passController = useTextEditingController();
     final confirmPassController = useTextEditingController();
     return SetupPage(
       icon: Icons.password,
-      title: context.locale.setupPasswordScreenTitle,
-      subtitle: context.locale.setupPasswordScreenDescription,
+      title: Text(context.locale.setupPasswordScreenTitle),
+      subtitle: Text(context.locale.setupPasswordScreenDescription),
       body: [
         Column(
           children: [
             TextFormField(
               decoration: InputDecoration(
                 contentPadding: const EdgeInsets.all(10),
-                errorText: error.value == '' ? null : error.value,
+                errorText: error.value,
                 errorStyle: const TextStyle(fontSize: 14),
                 labelText: locales.password,
                 border: const OutlineInputBorder(),
-                suffix: InkWell(
-                  onTap: () {
-                    if (obscure.value) {
-                      obscure.value = false;
-                    } else {
-                      obscure.value = true;
-                    }
-                  },
-                  child: Padding(
-                    padding: const EdgeInsets.only(right: 10.0),
-                    child: Icon(
-                      obscure.value ? Icons.visibility_off : Icons.visibility,
+                suffix: Padding(
+                  padding: const EdgeInsets.all(5),
+                  child: SizedBox(
+                    height: 20,
+                    width: 20,
+                    child: IconButton(
+                      iconSize: 25,
+                      // ignore: use_named_constants
+                      padding: const EdgeInsets.all(0),
+                      color: context.colorScheme.primary,
+                      icon: obscure.value
+                          ? const Icon(
+                              Icons.visibility,
+                            )
+                          : const Icon(
+                              Icons.visibility_off,
+                            ),
+                      onPressed: () => obscure.value = !obscure.value,
                     ),
                   ),
                 ),
@@ -62,18 +67,24 @@ class SetupPassword extends HookWidget {
                 errorStyle: const TextStyle(fontSize: 14),
                 labelText: locales.confirmPassword,
                 border: const OutlineInputBorder(),
-                suffix: InkWell(
-                  onTap: () {
-                    if (obscure.value) {
-                      obscure.value = false;
-                    } else {
-                      obscure.value = true;
-                    }
-                  },
-                  child: Padding(
-                    padding: const EdgeInsets.only(right: 10.0),
-                    child: Icon(
-                      obscure.value ? Icons.visibility_off : Icons.visibility,
+                suffix: Padding(
+                  padding: const EdgeInsets.all(5),
+                  child: SizedBox(
+                    height: 20,
+                    width: 20,
+                    child: IconButton(
+                      iconSize: 25,
+                      // ignore: use_named_constants
+                      padding: const EdgeInsets.all(0),
+                      color: context.colorScheme.primary,
+                      icon: obscure.value
+                          ? const Icon(
+                              Icons.visibility,
+                            )
+                          : const Icon(
+                              Icons.visibility_off,
+                            ),
+                      onPressed: () => obscure.value = !obscure.value,
                     ),
                   ),
                 ),
@@ -89,8 +100,8 @@ class SetupPassword extends HookWidget {
           ],
         ),
       ],
-      action: () async {
-        return await Core.auth.signUp(
+      action: () {
+        Core.auth.signUp(
           email: email,
           password: passController.text,
           confirmPassword: confirmPassController.text,
@@ -100,7 +111,6 @@ class SetupPassword extends HookWidget {
           context: context,
         );
       },
-      nextRoute: const SetupVerification(),
     );
   }
 }
