@@ -1,6 +1,5 @@
 import 'package:allo/logic/backend/general/users.dart';
 
 class General {
-  const General();
-  final UserMethods user = const UserMethods();
+  final UserMethods user = UserMethods();
 }
